chore: rename lib in tests to utils
This solves a compiler warning
